Add files via upload
In this project, I tried to look at the relationships between sales and customers' demographic, as well as customers' preferences. 
The analysis quickly and evidently showed that a host of the customers come from the United States, while the least customers come from the United Kingdom. 
Also, the visuals reveal that customers tastes were changing over the years for reasons not revealed in the chart, unfortunately. 

There are a lot of information to be gathered from the visuals, but just a hgihlight.